automated-matches-scraper/requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
pandas==2.0.3
google-cloud-storage==2.14.0
google-cloud-compute==1.14.0
webdriver-manager==4.0.1
python-dateutil==2.8.2
numpy==1.24.3
requests==2.31.0
lxml==4.9.3
urllib3==2.0.7
setuptools==69.2.0
packaging==24.0 